Clean-up dependencies and move real_time_correlative_scan_matcher.
