+++ cartographer/mapping/internal/2d/scan_matching/real_time_correlative_scan_matcher_2d_test.cc
@@ -24,6 +24,7 @@
 #include "cartographer/common/make_unique.h"
 #include "cartographer/mapping/2d/probability_grid.h"
 #include "cartographer/mapping/2d/range_data_inserter_2d.h"
+#include "cartographer/mapping/internal/scan_matching/real_time_correlative_scan_matcher.h"
 #include "cartographer/sensor/point_cloud.h"
 #include "cartographer/transform/transform.h"
 #include "gtest/gtest.h"
